Extract ConsumerInfo and SqlAsyncCallbackDelegate and migrate usages
