+++ src/Microsoft.Data.SqlClient/netcore/src/Interop/SNINativeMethodWrapper.Windows.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Runtime.InteropServices;
 using System.Text;
+using Interop.Windows.Sni;
 using Microsoft.Data.Common;
 using Microsoft.Data.SqlClient;
 
@@ -15,9 +16,6 @@ namespace Microsoft.Data.SqlClient
         private static int s_sniMaxComposedSpnLength = -1;
 
         private const int SniOpenTimeOut = -1; // infinite
-
-        [UnmanagedFunctionPointer(CallingConvention.StdCall)]
-        internal delegate void SqlAsyncCallbackDelegate(IntPtr m_ConsKey, IntPtr pPacket, uint dwError);
 
         internal delegate IntPtr SqlClientCertificateDelegate(IntPtr pCallbackContext);
 
@@ -36,14 +34,6 @@ namespace Microsoft.Data.SqlClient
         }
 
         #region Structs\Enums
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct ConsumerInfo
-        {
-            internal int defaultBufferSize;
-            internal SqlAsyncCallbackDelegate readDelegate;
-            internal SqlAsyncCallbackDelegate writeDelegate;
-            internal IntPtr key;
-        }
 
         [StructLayout(LayoutKind.Sequential)]
         internal struct AuthProviderInfo
